politeiad: Organize v1 routes.
